#21 Refactor auth source by splitting code into user_manager
